refactor: split wiki/servicecrudtest.refactor
